# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HeatZone AI – Uttar Pradesh Urban Heat Intelligence System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface City {
  id: number;
  name: string;
  latitude: number;
  longitude: number;
  population: number;
  populationDensity: number;
  totalArea: number;
  builtUpArea: number;
  industrialArea: number;
  residentialArea: number;
  roadArea: number;
  openLand: number;
  forestCover: number;
  urbanGreenSpace: number;
  treeDensity: number;
  ndvi: number;
  totalVehicles: number;
  petrolVehicles: number;
  dieselVehicles: number;
  electricVehicles: number;
  cngVehicles: number;
  createdAt: string;
  updatedAt: string;
}

export interface WeatherData {
  id: number;
  cityId: number;
  cityName: string;
  temperature: number;
  feelsLike: number;
  humidity: number;
  windSpeed: number;
  pressure: number;
  cloudCover: number;
  rainfall: number;
  /** @nullable */
  uvIndex?: number | null;
  weatherMain: string;
  weatherDescription: string;
  recordedAt: string;
}

export type HeatPredictionHeatZone =
  (typeof HeatPredictionHeatZone)[keyof typeof HeatPredictionHeatZone];

export const HeatPredictionHeatZone = {
  cool: "cool",
  moderate: "moderate",
  high: "high",
  extreme: "extreme",
} as const;

export interface HeatPrediction {
  id: number;
  cityId: number;
  cityName: string;
  heatRiskScore: number;
  heatZone: HeatPredictionHeatZone;
  temperature: number;
  humidity: number;
  vehicleDensity: number;
  populationDensity: number;
  greenCoverRatio: number;
  builtUpRatio: number;
  coolingIndex: number;
  trafficHeatFactor: number;
  latitude: number;
  longitude: number;
  predictedAt: string;
}

export type RecommendationPriority =
  (typeof RecommendationPriority)[keyof typeof RecommendationPriority];

export const RecommendationPriority = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface Recommendation {
  id: number;
  cityId: number;
  category: string;
  title: string;
  description: string;
  priority: RecommendationPriority;
  impact: string;
  icon: string;
  createdAt: string;
}

export interface CityDataset {
  city: City;
  latestWeather?: WeatherData;
  latestPrediction?: HeatPrediction;
  recommendations: Recommendation[];
  weatherHistory: WeatherData[];
  heatHistory: HeatPrediction[];
}

export interface DashboardOverview {
  totalCities: number;
  avgHeatRisk: number;
  extremeHeatCities: number;
  highHeatCities: number;
  moderateHeatCities: number;
  coolCities: number;
  avgTemperature: number;
  avgHumidity: number;
  totalVehicles: number;
  avgGreenCover: number;
  avgNDVI: number;
  avgNDBI: number;
  avgEmissionIndex: number;
  lastUpdated: string;
  cityPredictions: HeatPrediction[];
}

export interface RefreshResult {
  success: boolean;
  citiesUpdated: number;
  message: string;
}

export type GetWeatherHistoryParams = {
  limit?: number;
};

export type GetHeatHistoryParams = {
  limit?: number;
};